tests: move _TempModule to testSupport module
