+++ tests/filters/filters_test.py
@@ -6,33 +6,8 @@ from ufo2ft.filters import (
 from fontTools.misc.py23 import SimpleNamespace
 from fontTools.misc.loggingTools import CapturingLogHandler
 
-import sys
-import types
 import pytest
-
-
-class _TempModule(object):
-    """Temporarily replace a module in sys.modules with an empty namespace"""
-    def __init__(self, mod_name):
-        self.mod_name = mod_name
-        self.module = types.ModuleType(mod_name)
-        self._saved_module = []
-
-    def __enter__(self):
-        mod_name = self.mod_name
-        try:
-            self._saved_module.append(sys.modules[mod_name])
-        except KeyError:
-            pass
-        sys.modules[mod_name] = self.module
-        return self
-
-    def __exit__(self, *args):
-        if self._saved_module:
-            sys.modules[self.mod_name] = self._saved_module[0]
-        else:
-            del sys.modules[self.mod_name]
-        self._saved_module = []
+from ..testSupport import _TempModule
 
 
 class FooBarFilter(BaseFilter):
